[4006] Refactor the importers
Extract common logic into a base class.
